Introduce test-module.py
This is a simple utility that uses the Avocado Job API to run custom
jobs based on one metadata file.  There is one workaround to a bug in
current Avocado (the reason why "spawner.podman.avocado_spawner_egg"
is set).

On the CI side, it uses a newer than released (post 101.0) version of
Avocado, given that it has features needed here.

Signed-off-by: Cleber Rosa <crosa@redhat.com>