pyproject.toml init
This commit creates pyproject.toml for configuration of packaging tools.
This file contains basic information about Autils for deployment to PyPi
and configuration for Autils build.

The Autils use setuptools as build backend to ensure compatibility with
Avocado. The build version is dynamic and based on git tags, it will
make easier to automate build process during testing with test.pypi.

Reference: #21
Signed-off-by: Jan Richter <jarichte@redhat.com>