Release workflow
This is an introduction to pre-release and release wokflows.

The pre-release will build the current main branch a upload it into TEST
PyPi repo to test out the current build.

The release workflow will generate new autils major or minor input based
on the release. Then it will do the build and upload it to PyPi.

Reference: #21
Signed-off-by: Jan Richter <jarichte@redhat.com>