Environment preparation for module testing
This change uses containerfile for building container images with the
local changes of autils for testing. This approach will help us to use
containers for running selftests in different environments.

Since autils uses only redhat based containers, we use on container file
for removing duplicities and the actual system version is imported
during the runtime. This solution supports multiple containerfiles in
case that we will add different types of systems.

The build procedure uses the `--no-cache` flag to ensure that we always
test the latest changes. As a downside to this is that the images needs
to be always built which increases the test runtime.

Signed-off-by: Jan Richter <jarichte@redhat.com>