Code coverage report
After this change, it is possible to run test_moduel.py with --coverage
argument, which will create print the coverage results to console and
create coverag.xml file.

It uses avocado features for running coverage.py in containers. The
coverage report is stored in `/tmp/autils_coverage` inside container and
then transferred to the host by  `podman cp` command.

Reference: #24
Signed-off-by: Jan Richter <jarichte@redhat.com>