Reorg module metadata and test directory to avoid namespace clashes
Because some tools, such as pylint, will mimic the import of Python
modules, they can get confused on the test directory structure, and
confuse it with the file containing the module implementation.

For instance, because there is currently:

   autils/archive/ar.py

And:

   tests/autils/archive/ar.py

Pylint may consider the later to be the former.  Because we don't need
the autils component in the tests, and because it's not expected that
this repo will contain anything other than module implementations
under the "autils" component, the simplest solution is to drop
it.

For consistency, this also drops it from the metadata directory.

Signed-off-by: Cleber Rosa <crosa@redhat.com>