Dynamic versioning - remove local version
Current dynamic versioning configuration uses latest local git commit
has to generate local version of aautils. This causes troubles in
pre-release testing because such version in not compatible with PEP 643
and can't be uploaded to test.pypi. This commit removes the commit hash
from version and adds dev tag if the build is not part of the release
process.

Reference:
https://github.com/avocado-framework/aautils/actions/runs/12123321653/job/33799201398#step:3:164
Signed-off-by: Jan Richter <jarichte@redhat.com>